+++ yandex/resource_yandex_mdb_postgresql_user.go
@@ -2,6 +2,7 @@ package yandex
 
 import (
 	"fmt"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"log"
 	"time"
 
@@ -62,7 +63,10 @@ func resourceYandexMDBPostgreSQLUser() *schema.Resource {
 				Type:     schema.TypeList,
 				Optional: true,
 				Computed: true,
-				Elem:     &schema.Schema{Type: schema.TypeString},
+				Elem: &schema.Schema{
+					Type:         schema.TypeString,
+					ValidateFunc: validation.StringIsNotEmpty,
+				},
 			},
 			// TODO change to permissions
 			"permission": {